Update Rules::Taxon/Product handling of invalid match policies
For `Rules::Product`: Raise an error for invalid match policies. This
was already happening in `Rules::Product#actionable?`, it just wasn't
happening in `Rules::Product#eligible`.

For `Rules::Taxon`: Log a warning and assume it was meant to be an
`any` match policy.  I'm being more lenient here because
`Rules::Taxon` had a bug with the default value for match policies and
assuming `any` is essentially what the previous logic did.

Also, add validations to help avoid invalid values in the first place.