Re-write slow Order/Shipment migration with SQL
This migration was originally written with SQL, but was updated to use
ActiveRecord logic due to an issue related to shipments with no cost:

https://github.com/spree/spree/commit/2b7638eeeb6c5c2f0dded0122ce9a82c1d0f6956

However, the performance of the ActiveRecord version of this query is
significantly impacted by large numbers of Orders and Shipments.

Fortunately for us, the previous SQL version only needs a `COALESCE()`
wrapper around the sub-select to address the same issue the AR version
was written to fix.