Invoke on payment method not payment
It seems to be convention that payment methods do business in cents, and
it was certainly the case that the store credit payment method #cancel
was expecting an amount in cents (which it prompty converts to a
decimal to pass down the store credit itself).

This worked before because we were incorrectly calling cancel on the
stopre credit directly (which expects dollars) but now that we're
correctly calling the store credit payment method we owe it cents.

Also the store credit currency was being passed through explicitly, but
now that's the job of store_credit_payment#credit, who performs the same
lookup and passes it along.