Allow customization of link_to_delete confirmation
When writing custom views for the admin, it's useful to be able to
provide action buttons in a similar way to Solidus' existing edit/delete
buttons. However, when creating delete buttons with `link_to_delete`,
it's currently not possible to provide a custom confirmation message.

This commit addresses that, by allowing the `confirm` option to be
provided and  used in place of the standard "Are you sure?" prompt.