Fix FreeShipping promos that have codes
This commit:
https://github.com/solidusio/solidus/commit/4a62eca#diff-d7fd5424af325ba4827da883ac185ded
broke this fix:
https://github.com/solidusio/solidus/commit/7b50d73
because we stopped looking for promotions with codes.  So if you apply
a free shipping promo to your order and then go backward in the order
state machine (and have your shipments regenerated) you won't get your
promotion re-applied when you move forward in the state machine again.

This also updates FreeShipping promotions to use the recently-added
apply_automatically promotion flag, rather than looking for the
absence of codes and paths.