Remove require_email on email validation
We only need `require_email` to be true to validate the presence of an
email address. If, for whatever reason, you want to add an email address
to an order *before* `require_email` is true then it should still go
through the appropriate validations to make sure it's a valid email
address.