Fix flaky specs
We are now triggering caching for shipments an line items when we call
`order.contents.add`. This meant that `order.shipments` and
`order.line_items` needed to be reloaded for these specs. Also, the
`order.shipments.reload` that I had added caused flakiness because
in Postgres they would sometimes reload in a random order and then
`shipments.first` and `shipments.last` wouldn't be the expected items.