Remove use of order.tax_zone in taxation code
This commit removes call to `Spree::Order#tax_zone` from the codebase.

These calls are problematic, because whenever we have to check whether
a zone includes another zone entirely, we have to iterate over all of
both zones members. Really, it's enough to check whether the order's tax
address is within the desired zone (for a taxation or shipping zone).

This also comes with a number of fixes in specs where calls to `order.tax_zone`
were abused for test setup.