Add link to store credit from payment page
Currently, when a store credit is consumed in a payment, there is no
link to that store credit record in either the order payments show page
or the order payments index page.

This makes it difficult to determine which particular store credit was
consumed, the history of that store credit, etc.

This commit addresses these concerns by adding a link back to the store
credit from the payment page itself.