Wrap OrderUpdater#update in a transaction
As of #1479 we are doing more work inside of OrderUpdater. We're
doing additions and deletions of tax adjustments, which also causes more
touching of records.

This is all made faster by wrapping the entire update in a transaction.
All the writes can be queued together by the DB, and as of rails 5,
dependent touching is grouped together to the end of the transaction.