Remove tax adjustments when no zone matches
Previously when the tax_address's country_id was nil, OrderAdjuster and
ItemAdjuster's #adjust! would exit without making any changes to tax
adjustments. If the tax address had a value, and then later becomes nil,
the order would "remember" the old taxes and not update them.

Fortunately, clearing the address from an order is an uncommon
operation. AFAIK it can only be done on the console.

This commit changes the item and order adjuster to continue in this
case so that the order will have taxes removed.