Mark tax_adjuster_class as @api experimental
This will hide tax_adjuster_class in our generated docs.

We're planning to make some changes to this API to reduce the work
needing to be done on the tax integration. We don't know right now if
this interface will be a good fit when we are done, but we would like to
include it in some way for now.