Improve editing of shipping calculators
Previously, after changing the type of shipping calculator, the user
would need to save the page before seeing or being able to edit any of
the values for that calculator.

This PR changes the calculator_fields partial to render all values for
all types of preferences, and then hide and disable (which prevents form
submission) the fields for calculator types other than the one selected.