Add failing specs on AvailabilityValidator
These specs exibit two issues when a line item is split across multiple
shipments.

If the line item is split across shipments, any availabilty errors will
be reported twice.

If a shipment is split across shipments with the same stock location, it
won't detect insufficient stock if both shipments are able to separately
fulfill the inventory.