Convert spree.js.coffee to plain JS
This file's history was already replaced by the previous commit, so we
might as well convert to plain JS at the same time.

This is the only CoffeeScript (now only JS) file in the solidus_core
gem, so it can now in theory be used in an app without needing to pull
in CoffeeScript (assuming the app doesn't include solidus_frontend or
soldius_backend).