Migrate all reusable credit card to wallet sources
The behaviour prior to adding the Wallet was to allow reuse of any
CreditCard which had a payment profile. However the migration was only
migrating CreditCards marked "default".

This now migrates all credit cards with an associated
gateway_customer_profile_id.

This also no longer removes the "default" column from the credit card,
as we decided this was safer to leave in for at least one release.